Per @gregmuellegger suggestion, I've implemented a new operation called AlterSortedManyToManyField, which extends AlterField.
To use this operation you have to replace it explicitly on the migration. Create the migration with "makemigrations" and then replace AlterField with this one.

This is only for Django 1.7 migrations. Didn't change or add anything for South.

TO-DO: add the default thingy successfully, like create_sorted_many_to_many_intermediate_model() does. Right now sort_value is created (or whatever it's named)
but no default is set.