Fixed performance when there is a lot of item in the m2m relation and a visual glitch.
Formerly the jQuery code was looping through all the potential <input> elements
that could be added to the m2m relationship in order to find the parent
<ul> elements. It is now done this other way around, we only loop
through <ul> elements that have <input> child with the `sortedm2m`
class.

We are also hidding the <ul> until it gets fully prepared, to avoid to have
a bad flicker when initially loading the page.