Use different function name for dismissing "add" popup when necessary
Django 1.8 introduced a change of the function name from
`dismissAddAnotherPopup` to `dismissAddRelatedObjectPopup`. Old name was
kept for backwards compatibility, but since it's value is now a
reference and not an actual function, sortedm2m overrides only that
reference and not the actual `dismissAddRelatedObjectPopup` which is
now the function being called in
`django/contrib/admin/templates/admin/popup_response.html`. Since django
no longer calls the overridden function, the added object would only
appear after refresh.

Ref: https://github.com/django/django/commit/07988744b347302925bc6cc66511e34224db55ab#diff-ec17958cc9f7c2542303557dfc4ad997R130